Add checking for related wagons before deleting train row.

+++ app/src/main/java/com/example/coursework/ui/train/viewModels/TrainDetailsViewModel.kt
@@ -4,14 +4,18 @@ import androidx.lifecycle.SavedStateHandle
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.example.coursework.data.repositories.TrainsRepository
+import com.example.coursework.data.repositories.relationsRepositories.TrainWithWagonsRepository
 import com.example.coursework.ui.state.*
 import com.example.coursework.ui.train.screens.TrainDetailsDestination
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.async
 import kotlinx.coroutines.flow.*
 
 
 class TrainDetailsViewModel(
     savedStateHandle: SavedStateHandle,
-    private val trainsRepository: TrainsRepository
+    private val trainsRepository: TrainsRepository,
+    private val trainWithWagonsRepository: TrainWithWagonsRepository
 ) : ViewModel() {
 
     private val trainId: Int = checkNotNull(savedStateHandle[TrainDetailsDestination.trainIdArg])
@@ -25,8 +29,18 @@ class TrainDetailsViewModel(
             initialValue = TrainUiState()
         )
 
-    suspend fun deleteTrain() {
-        trainsRepository.deleteTrain(uiState.value.toTrain())
+    suspend fun validateTrainDeletion(): String {
+        val message = viewModelScope.async(Dispatchers.IO) {
+            val trainWithWagonsList =
+                trainWithWagonsRepository.getTrainsAndWagons()
+            if (trainWithWagonsList.any { uiState.value.id == it.train.id && it.wagons.isNotEmpty() }) {
+                "This train can't be deleted because it's used in existing wagon(-s)."
+            } else {
+                trainsRepository.deleteTrain(uiState.value.toTrain())
+                "Row deleted successfully."
+            }
+        }
+        return message.await()
     }
 
     companion object {
